chore!: miscellaneous cleanup (#16)
* Include readme in crate documetation

* Changed index-like numbers to usize

* BREAKING CHANGE: input intnums are now a hash map

* Minor code cleanup

* More minor cleanup

* Made pool size configurable

* Reuse the same pools for all functions

+++ src/polynomial/prettyprint.rs
@@ -27,7 +27,7 @@ impl<'a, T: fmt::Display> fmt::Display for PrettyPrintPolynomial<'a, T> {
                 .properties
                 .semigroup
                 .elements
-                .column(*i as usize)
+                .column(*i)
                 .iter()
                 .enumerate()
             {
